[22/04/08] c++ core guidelines. #9
F21. To return multiple "out" values, prefer returning a struct or tuple
F22. Use T* or owner<T*> to designate a single object
F23. Use a not_null<T> to indicate that "null" is not a valid value
F24. Use a span<T> or a span_p<T> to designate a half-open sequence
F25. Use a zstring or a not_null<zstring> to designate a C-sytle string
F26. Use a unuque_ptr<T> to transfer ownership where a pointer is needed
F27. Use a shared_ptr<T> to share ownership