[22/04/11] c++ core guidelines. #12
C1. Organize related data into structures (sturcts or classes)
C2. Use class if the class has an invariant; use struct if the data members can vary independently
C3. Represent the distinction between an interface and an implementation using a class
C4. Make a function a member only if it needs direct access to the representation of a class
C5. Place helper functions in the same namespace as the class they support
C7. Don't define a class or enum and declare a variable of its type in the same statement
C8. Use class rather than struct if any member is non-public
C9. Minimize exposure of members