[22/04/13] c++ core guidelines. #14
C30. Define a destructor if a class needs an explicit action at object destruction
C31. All resources acquired by a class must be released by the class's destructor
C32. If a class has a raw pointer (T*) or reference (T&), consider whether it might be owning
C33. If a class has an owning pointer member, define a destructor
C35. A base class destructor should be either public and virtual, or protected and non-virtual