[22/04/22] c++ core guidelines. #22
C100. Follow the STL when defining a container
C101. Give a container value semantics
C102. Give a container move operators
C103. Give a container an initializer list constructor
C104. Give a container a default constructor that sets it to empty
C109. If a resource handler has pointer semantics, provide * and ->