[22/04/23] c++ core guidelines. #23
C120. Use class hierarchies to represent concepts with inherent hierarchical structure
C121. If a base class is used as an interface, make it a pure abstract class
C122. Use abstract classes as interfaces when complete separation of interface and implementation is needed
C126. An abstract class typically doesn't need a user-written constructor
C127. A class with a virtual function should have a virtual or protected destructor